Check directories before building

+++ Makefile
@@ -1,5 +1,6 @@
 BOWER			= bower
 CP				= cp -r
+DIR_EXISTS		= test -d
 GRUNT			= grunt
 MKDIR			= mkdir
 NPM				= npm
@@ -9,6 +10,7 @@ SEP				= /
 
 ifeq ($(OS), Windows_NT)
 CP				= XCOPY /S
+DIR_EXISTS		= exist
 RM				= RD /S /Q
 DIRTY_SEP		= \ 
 SEP				= $(strip $(DIRTY_SEP))
@@ -34,7 +36,8 @@ $(INTERFACE_DEPS):
 
 interface: $(INTERFACE_DEPS)
 	cd $(INTERFACE_ROOT) && $(GRUNT) build
-	$(MKDIR) $(INTERFACE_OUT) && $(CP) $(INTERFACE_ROOT)$(SEP)dist$(SEP)* $(INTERFACE_OUT)
+	$(DIR_EXISTS) $(INTERFACE_OUT) || $(MKDIR) $(INTERFACE_OUT)
+	$(CP) $(INTERFACE_ROOT)$(SEP)dist$(SEP)* $(INTERFACE_OUT)
 
 baasbox: $(BAASBOX_OUT)
 
@@ -46,8 +49,9 @@ clean:
 	cd $(BAASBOX_ROOT) && $(PLAY) clean
 
 mrproper: clean
-	$(RM) $(INTERFACE_DEPS)
-	$(RM) $(INTERFACE_OUT)
-	$(RM) $(BAASBOX_OUT)
+	$(DIR_EXISTS) $(INTERFACE_ROOT)$(SEP)node_modules && $(RM) $(INTERFACE_ROOT)$(SEP)node_modules
+	$(DIR_EXISTS) $(INTERFACE_ROOT)$(SEP)app$(SEP)bower_components && $(RM) $(INTERFACE_ROOT)$(SEP)app$(SEP)bower_components
+	$(DIR_EXISTS) $(INTERFACE_OUT) && $(RM) $(INTERFACE_OUT)
+	$(DIR_EXISTS) $(BAASBOX_OUT) && $(RM) $(BAASBOX_OUT)
 
 .PHONY: baasbox clean interface mrproper tomat
